+++ test/body-parser.js
@@ -1,4 +1,6 @@
 
+process.env.NO_DEPRECATION = 'body-parser'
+
 var assert = require('assert');
 var http = require('http');
 var request = require('supertest');